Chunk index issue fixed. That problem was being caused becouse Javascript object files transform all keys in string by default. That way, when Object.entries was called all indexes used were strings, causing problems.
To avoid this problem, all objects that uses chunk indexes as keys were
changed to Maps.

To improve performace, the blobfile_join worker was changed to base64_datajoin where intead of create and updating blob files, all the raw bytes are saved in memory. In the end of transmission other routine should create a blob file from this data.